+++ package.json
@@ -14,7 +14,6 @@
   },
   "dependencies": {
     "@react-native-community/masked-view": "0.1.6",
-
     "@react-navigation/material-bottom-tabs": "^5.2.9",
     "@react-navigation/native": "^5.1.1",
     "@react-navigation/routers": "^5.3.0",
