+++ src/Navigation/Navigation.js
@@ -12,7 +12,10 @@ import TrainingPage from '../Components/TrainingPage';
 import ProgramListPage from '../Components/ProgramListPage';
 import ProgramPage from '../Components/ProgramPage';
 import AnimationEditorPage from '../Components/AnimationEditorPage';
-import TheoryPage from '../Components/TheoryPage';
+
+import EssentialPage from '../Components/EssentialPage';
+import Dictionary from '../Components/Dictionary';
+import TacticsPage from '../Components/TacticsPage';
 
 import I18n from '../utils/i18n';
 
@@ -71,12 +74,8 @@ export const Navigation = () => (
       component={AnimationEditorPage}
       options={() => ({ title: I18n.t('navigation.animationEditorPage') })}
     />
-    <Stack.Screen
-      name="TheoryPage"
-      component={TheoryPage}
-      options={() => ({
-        title: 'Theory',
-      })}
-    />
+    <Stack.Screen name="Dictionary" component={Dictionary} options={{ title: I18n.t('navigation.dictionary') }} />
+    <Stack.Screen name="EssentialPage" component={EssentialPage} options={{ title: I18n.t('navigation.essential') }} />
+    <Stack.Screen name="TacticsPage" component={TacticsPage} options={{ title: I18n.t('navigation.tactics') }} />
   </Stack.Navigator>
 );
