+++ src/utils/locales/en.js
@@ -51,16 +51,23 @@ export default {
     programListPage: 'Choose a program',
     drillPageMinimal: '{{training}} drills',
     animationEditorPage: 'New Drill',
+    dictionary: 'Dictionary',
+    essential: 'Essential',
+    tactics: 'Tactics',
   },
   homePage: {
-    frisbeeTab: 'Frisbee',
+    trainingTab: 'Training',
     fitnessTab: 'Fitness',
-    theory: 'Theory',
+    theoryTab: 'Theory',
+    editor: 'Editor',
     drills: 'Drills',
+    dictionary: 'Dictionary',
+    essential: 'Essential',
+    tactics: 'Tactics',
     programs: 'Practice Programs',
     leanTitle: 'Lean Fit',
     leanSubtitle: 'Get lean and fit over 6 weeks with a balanced plan that builds endurance',
-    bodyweightTitle: 'Bodywaight only',
+    bodyweightTitle: 'Bodyweight only',
     bodyweightSubtitle: 'Push your strenght and improve muscle tone over the season - all without weights',
     gymTitle: 'Gym Strong',
     gymSubtitle: 'Build full-body strenght with a focus on weight training over the season',
