+++ src/utils/locales/fr.js
@@ -51,12 +51,18 @@ export default {
     programListPage: "Programmes d'entraînement",
     drillPageMinimal: '{{training}} - exercices',
     animationEditorPage: 'Nouvel exo',
+    dictionary: 'Dictionnaire',
+    essential: 'Essentiel',
+    tactics: 'Tactiques',
   },
   homePage: {
     frisbeeTab: 'Frisbee',
     fitnessTab: 'Fitness',
     theory: 'Théorie',
     drills: 'Exercices',
+    dictionary: 'Dictionnaire',
+    essential: 'Essentiel',
+    tactics: 'Tactiques',
     programs: "Programmes d'entraînement",
     leanTitle: 'Affuté',
     leanSubtitle: 'Préparez-vous en 6 semaines avec un programme équilibré pour votre endurance',
